heap_now returns usize: that is what the bindings say
The probe declared `-> (u32, u32)`. `heap_caps_get_free_size` and
`heap_caps_get_largest_free_block` both return `usize`:

    pub fn heap_caps_get_free_size(caps: u32) -> usize;
    pub fn heap_caps_get_largest_free_block(caps: u32) -> usize;

(esp-idf-sys generated bindings, riscv32imc-esp-espidf.)

`log_memory_profile` never had to name the type because it only
interpolates the values into a format string, so the mismatch appeared
the moment a function returned them. Checked against the generated
bindings this time rather than inferred.

+++ src/metrics.rs
@@ -155,7 +155,7 @@ pub mod alloc_note {
 /// that prints four lines per call pushes the evidence out of it. That is not
 /// hypothetical -- an earlier `memory_report()` probe in this investigation
 /// answered its question and displaced the run-up to the crash doing it.
-pub fn heap_now() -> (u32, u32) {
+pub fn heap_now() -> (usize, usize) {
     // SAFETY: read-only ESP-IDF accessors, safe from any context.
     let cap = sys::MALLOC_CAP_INTERNAL | sys::MALLOC_CAP_8BIT;
     unsafe {
